+++ src/main/java/frc/robot/Constants.java
@@ -26,10 +26,6 @@ public final class Constants {
         public static final int A_BUTTON = 1, B_BUTTON = 2, X_BUTTON = 3, Y_BUTTON = 4;
     }
     public static class MovementConstants {
-        public static final double CLIMBER_MOTOR_SPEED = 0.5;
         public static final int CLIMBER_MAX_HEIGHT = 100; // temporary until physical measurements can be made
-        public static final int CLIMBER_MIN_HEIGHT = 0; // likewise
-        public static final double NEO_ENCODER_RADIUS = 2; // likewise
-        public static final double SPARK_ENCODER_RADIUS = 2; // likewise
     }
 }
