+++ src/main/java/frc/robot/commands/ClimberPID.java
@@ -8,21 +8,21 @@ import edu.wpi.first.math.trajectory.TrapezoidProfile;
 import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 
-public class ClimberPIDTest extends CommandBase {
+public class ClimberPID extends CommandBase {
     private Climber climber;
     private Joystick joy;
     private double kP = 0.0;
     private double kD = 0.0;
     private double kI = 0.0;
 
-    public ClimberPIDTest(Climber m_climber, Joystick joystick) { 
+    public ClimberPID(Climber m_climber, Joystick joystick) { 
         addRequirements(m_climber);
         climber = m_climber;
         joy = joystick;
     }
     private final TrapezoidProfile.Constraints m_constraints = new TrapezoidProfile.Constraints(1.75, 0.75);
     private final ProfiledPIDController m_controller = new ProfiledPIDController(0.01, 0.0, 0.0, m_constraints);
-    private double targetHeight = 0;
+    private static double targetHeight = 0;
     private double climberMotorDistance;
     private double PIDOutput;
 
@@ -33,16 +33,16 @@ public class ClimberPIDTest extends CommandBase {
     }
 
     @Override
-    public void initialize() {}
-
-    @Override
-    public void execute() {
-
+    public void initialize() {
         kP = SmartDashboard.getNumber("kP", 0.01);
         kI = SmartDashboard.getNumber("kI", 0);
         kD = SmartDashboard.getNumber("kD", 0);
 
         m_controller.setPID(kP, kI, kD);
+    }
+
+    @Override
+    public void execute() {
 
         if (-1*joy.getRawAxis(Constants.Controller.JOYSTICK_1) > 0.1) {
             targetHeight += 0.1;
